@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Text.Json.Serialization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
@@ -25,18 +27,17 @@ namespace Playing_with_Serilog
 
     public void ConfigureServices(IServiceCollection services)
     {
-      services.AddMvc(options =>
-      {
-        options.Filters.Add<ExecutionTimeLogFilter>(); // Add filter
-      })
-      .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+      services
+        .AddControllers(options => options.Filters.Add<ExecutionTimeLogFilter>())
+        .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
+        .AddJsonOptions(options => options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter()));
 
       services.AddSingleton(_loggingLevelSwitch);
 
       services.AddTransient<ExecutionTimeLogFilter>();
     }
 
-    public void Configure(IApplicationBuilder app, IHostingEnvironment env, IApplicationLifetime appLifetime)
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IHostApplicationLifetime appLifetime)
     {
       if (env.IsDevelopment())
         app.UseDeveloperExceptionPage();
@@ -44,10 +45,12 @@ namespace Playing_with_Serilog
       // Make sure to flush the log.
       appLifetime.ApplicationStopped.Register(Log.CloseAndFlush);
 
+      app.UseRouting();
+
       // Request logging is a new feature in the Serilog.AspNetCore 3.0.0
       app.UseSerilogRequestLogging();
 
-      app.UseMvc();
+      app.UseEndpoints(endpoints => endpoints.MapControllers());
     }
 
     private void initLogger()
@@ -69,7 +72,7 @@ namespace Playing_with_Serilog
       //  .MinimumLevel.ControlledBy(_loggingLevelSwitch)
       //  .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
       //  .MinimumLevel.Override("System", LogEventLevel.Warning)
-      //  .MinimumLevel.Override("Microsoft.AspNetCore.Hosting.Internal.WebHost", LogEventLevel.Information)
+      //  .MinimumLevel.Override("Microsoft.Hosting", LogEventLevel.Information)
       //  .Enrich.FromLogContext()
       //  .Enrich.WithProperty("EnrichProperty", "value")
       //  .WriteTo.Console(outputTemplate: "[{Timestamp:HH:mm:ss} {Level}] {Message}{NewLine}{Exception}")
